+++ src/main/java/com/example/demo/model/criteria/IncomeMoneyCriteria.java
@@ -1,5 +1,6 @@
 package com.example.demo.model.criteria;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,7 +10,7 @@ public class IncomeMoneyCriteria {
   private String sourceOrganization;
   private String invoiceReference;
   private String description;
-  private Integer amount;
+  private BigDecimal amount;
   private String jobId;
   private String incomeTypeId;
   private Boolean moneyReceived;
